requirements.txt
Flask
gspread
pandas
oauth2client
groq
markdown2
python-dotenv